+++ nano/node/CMakeLists.txt
@@ -157,6 +157,8 @@ add_library(
   telemetry.cpp
   transport/channel.hpp
   transport/channel.cpp
+  transport/channel_tcp.hpp
+  transport/channel_tcp.cpp
   transport/fake.hpp
   transport/fake.cpp
   transport/inproc.hpp
@@ -165,8 +167,8 @@ add_library(
   transport/message_deserializer.cpp
   transport/socket.hpp
   transport/socket.cpp
-  transport/tcp.hpp
-  transport/tcp.cpp
+  transport/tcp_channels.hpp
+  transport/tcp_channels.cpp
   transport/tcp_listener.hpp
   transport/tcp_listener.cpp
   transport/tcp_server.hpp
