+++ nano/node/bootstrap/bootstrap_frontier.cpp
@@ -2,7 +2,6 @@
 #include <nano/node/bootstrap/bootstrap_frontier.hpp>
 #include <nano/node/bootstrap/bootstrap_legacy.hpp>
 #include <nano/node/node.hpp>
-#include <nano/node/transport/tcp.hpp>
 #include <nano/secure/ledger.hpp>
 #include <nano/secure/ledger_set_any.hpp>
