+++ nano/node/transport/channel_tcp.cpp
@@ -1,0 +1,101 @@
+#include <nano/lib/stats.hpp>
+#include <nano/lib/utility.hpp>
+#include <nano/node/node.hpp>
+#include <nano/node/transport/channel_tcp.hpp>
+#include <nano/node/transport/message_deserializer.hpp>
+
+/*
+ * channel_tcp
+ */
+
+nano::transport::channel_tcp::channel_tcp (nano::node & node_a, std::weak_ptr<nano::transport::socket> socket_a) :
+	channel (node_a),
+	socket (std::move (socket_a))
+{
+}
+
+nano::transport::channel_tcp::~channel_tcp ()
+{
+	nano::lock_guard<nano::mutex> lk{ channel_mutex };
+	// Close socket. Exception: socket is used by tcp_server
+	if (auto socket_l = socket.lock ())
+	{
+		socket_l->close ();
+	}
+}
+
+void nano::transport::channel_tcp::update_endpoints ()
+{
+	nano::lock_guard<nano::mutex> lk (channel_mutex);
+
+	debug_assert (endpoint == nano::endpoint{}); // Not initialized endpoint value
+	debug_assert (local_endpoint == nano::endpoint{}); // Not initialized endpoint value
+
+	if (auto socket_l = socket.lock ())
+	{
+		endpoint = socket_l->remote_endpoint ();
+		local_endpoint = socket_l->local_endpoint ();
+	}
+}
+
+void nano::transport::channel_tcp::send_buffer (nano::shared_const_buffer const & buffer_a, std::function<void (boost::system::error_code const &, std::size_t)> const & callback_a, nano::transport::buffer_drop_policy policy_a, nano::transport::traffic_type traffic_type)
+{
+	if (auto socket_l = socket.lock ())
+	{
+		if (!socket_l->max (traffic_type) || (policy_a == nano::transport::buffer_drop_policy::no_socket_drop && !socket_l->full (traffic_type)))
+		{
+			socket_l->async_write (
+			buffer_a, [this_s = shared_from_this (), endpoint_a = socket_l->remote_endpoint (), node = std::weak_ptr<nano::node>{ node.shared () }, callback_a] (boost::system::error_code const & ec, std::size_t size_a) {
+				if (auto node_l = node.lock ())
+				{
+					if (!ec)
+					{
+						this_s->set_last_packet_sent (std::chrono::steady_clock::now ());
+					}
+					if (ec == boost::system::errc::host_unreachable)
+					{
+						node_l->stats.inc (nano::stat::type::error, nano::stat::detail::unreachable_host, nano::stat::dir::out);
+					}
+					if (callback_a)
+					{
+						callback_a (ec, size_a);
+					}
+				}
+			},
+			traffic_type);
+		}
+		else
+		{
+			if (policy_a == nano::transport::buffer_drop_policy::no_socket_drop)
+			{
+				node.stats.inc (nano::stat::type::tcp, nano::stat::detail::tcp_write_no_socket_drop, nano::stat::dir::out);
+			}
+			else
+			{
+				node.stats.inc (nano::stat::type::tcp, nano::stat::detail::tcp_write_drop, nano::stat::dir::out);
+			}
+			if (callback_a)
+			{
+				callback_a (boost::system::errc::make_error_code (boost::system::errc::no_buffer_space), 0);
+			}
+		}
+	}
+	else if (callback_a)
+	{
+		node.background ([callback_a] () {
+			callback_a (boost::system::errc::make_error_code (boost::system::errc::not_supported), 0);
+		});
+	}
+}
+
+std::string nano::transport::channel_tcp::to_string () const
+{
+	return nano::util::to_str (get_tcp_endpoint ());
+}
+
+void nano::transport::channel_tcp::operator() (nano::object_stream & obs) const
+{
+	nano::transport::channel::operator() (obs); // Write common data
+
+	obs.write ("socket", socket);
+}
