+++ nano/node/transport/fwd.hpp
@@ -3,4 +3,6 @@
 namespace nano::transport
 {
 class channel;
+class channel_tcp;
+class tcp_channels;
 }
